+++ changelog.md
@@ -3,6 +3,8 @@
 - added the Galaxy Soul
 - fixed number formatting, now they comma seperate per 000 so 1,000 etc.
 - added special item type
+- added blacksmith items type
+- added iridium ingot
 
 ## [0.0.18]
 ### Changes